don't try to start tunneler in intercept mode unless on linux
