+++ tunnel/dns/dns_notlinux.go
@@ -1,0 +1,11 @@
+//go:build !linux
+
+package dns
+
+func NewDnsServer(addr string) (Resolver, error) {
+	return nil, nil
+}
+
+func flushDnsCaches() {
+	// not implemented
+}
